flask_server/requirements.txt
aiofiles==22.1.0
aioflask==0.4.0
black==22.10.0
blinker==1.5
click==8.1.3
Flask==2.2.2
greenlet==2.0.0
greenletio==0.9.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
priority==2.0.0
python-dotenv==0.21.0
pytube==12.1.0
quart==0.18.3
six==1.16.0
toml==0.10.2
tomli==2.0.1
uvicorn==0.19.0
Werkzeug==2.2.2
wsproto==1.2.0
